drop unused Pledge._email, CancelledPledge._email fields now that we've migrated to AnonymousUser objects
I noticed there was an error in the migration. Canceled anonymous pledges got assigned a new AnonymousUser with the email address set according to the wrong variable. In the production database that variable was luckily None, and it's low value, so I'm not going to fix that. I'm updating the CancelledPledge admin to properly indicate what's going on.