make IncompletePledge unique on email and update get_or_create to only use the unique column in the get-part
We were getting 'contrib.models.MultipleObjectsReturned: get() returned more than one IncompletePledge -- it returned 2!' in get_or_create, because of a likely race condition. This will prevent an error during get (which is persistent--once the duplicate record goes in, the get is broken), although it shifts the problem to create (but only in a race condition, and the duplicate record will never go in).