update to Django 1.9 and bump HTML email dependencies
The pynliner branch that we were using for Py3 became incompatible with bs4. Py3 support was merged into their main line, so we can now install through a pip dependency of django-html-emailer, rather than manually.

Also dropping the submodule for django-html-emailer and switching to installing it via pip with a git+http reference. That brings in its pynliner dependency, but not its new commonmark-py-plaintext dependency which I've now added to our setup.

But the new packages required Django 1.9, so we're upgrading that too. I don't see anything in our code that needs updating.