simplify party filters by testing on the party stored on the Recipient object
For challengers the logic is unchanged. For incumbents, filtering is now on the actor's current party rather than on their party at the time of the action. Obviously this almost never matters. I think this is preferable anyway, but mostly I'm just doing this to simplify the code.

create_actors now sets the party field on incumbent recipients to the incumbent's current party - this was missing before. Test fixture updated.